// src/app/models/zscore.model.ts
export class ZScore {
    id: number;
    length: number;
    weight: number;
    months: number;
    zScoreType: string;
    gender: string;
    resultCode: string;

    createdBy: string;
    createdDateTime: Date;
    modifiedBy: string;
    modifiedDateTime: Date | null;
    status: boolean ;
}